backend/requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary
pydantic-settings
PyJWT
passlib[bcrypt]
fastapi-mail
websockets